trying to make the training run in GPU

+++ event/event_processor.py
@@ -265,6 +265,7 @@ class EventProcessor:
 
 
 def get_read_grad(vector: jnp.array):
+    vector = vector.astype('float64')
     grad = jacrev(EventProcessor.write_process)(vector)
     # [12 * 4 * 4]
     return grad[:, 1, :]
@@ -281,7 +282,7 @@ def get_write_grad(vector: jnp.array
 
 
 def get_create_grad(vector: jnp.array):
-    
+    vector = vector.astype('float64')
     grad = jacrev(EventProcessor.write_process)(vector)
     # [12 * 4 * 4]
     return grad[:, 1, :]
@@ -289,6 +290,7 @@ def get_create_grad(vector: jnp.array):
 
 
 def get_load_grad(vector: jnp.array):
+    vector = vector.astype('float64')
     grad = jacrev(EventProcessor.write_process)(vector)
     # [12 * 4 * 4]
     return grad[:, 1, :]
@@ -296,6 +298,7 @@ def get_load_grad(vector: jnp.array):
 
 
 def get_exec_grad(vector: jnp.array):
+    vector = vector.astype('float64')
     grad = jacrev(EventProcessor.write_process)(vector)
     # [12 * 4 * 4]
     return grad[:, 1, :]
